requirements.txt
# Confluent Kafka
confluent-kafka[avro,json,protobuf]==2.3.0
confluent-kafka[schema-registry]==2.3.0

# Google Cloud / Vertex AI
google-cloud-aiplatform==1.38.1
google-cloud-storage==2.14.0
google-cloud-bigquery==3.14.1
google-auth==2.25.2

# Data Processing & ML
numpy==1.24.3
scipy==1.11.4
pandas==2.1.4
scikit-learn==1.3.2

# Signal Processing
pywavelets==1.5.0

# API & Web
fastapi==0.108.0
uvicorn[standard]==0.25.0
requests==2.31.0
python-dotenv==1.0.0

# Slack Integration
slack-sdk==3.26.1

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
faker==21.0.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
